The unit for the hourly demand data is MW rather than kW
I checked with U.S. EIA data and Matt Shaner's processing code. The unit for the hourly demand data is MW rather than kW. As a result, I have updated the unit for energy quantities and capacities in this output script. If you use the normalized demands, it won't be an issue.